requirements-dev.txt
pytest==7.4.3
ruff==0.1.6
pre-commit==3.5.0
requests-mock==1.11.0
